Refactor and update family-related features and tests
- Add new family and family member controllers, models, schemas, and services
- Remove activity-related files and references
- Update user-related models, services, and tests
- Add new migration scripts and configuration files
- Update existing tests and add new tests for family-related features
- Modify database and routes to accommodate new changes
- Update web hooks server script

+++ api/run_migrations.sh
@@ -1,4 +1,17 @@
-env_vars=$(grep -v '^#' .env.staging | grep -v '^[[:space:]]*$' | sed 's/^/-e /' | tr '\n' ' ')
-cmd="docker run --rm -it $env_vars gcr.io/serene-smoke-429712-n5/family-planner-api alembic upgrade head"
-echo $cmd
-eval $cmd
+#if $1 is not provided, use staging
+stage=$1
+if [ -z "$stage" ]; then
+    env_file=".env"
+else
+    env_file=".env.$stage"
+fi
+if [ "$stage" == "staging" ]; then  # Updated comparison operator
+    env_vars=$(grep -v '^#' $env_file | grep -v '^[[:space:]]*$' | sed 's/^/-e /' | tr '\n' ' ')
+    cmd="docker run --rm -it $env_vars gcr.io/serene-smoke-429712-n5/family-planner-api alembic upgrade head"
+    echo $cmd
+    eval $cmd
+else
+    export $(grep -v '^#' $env_file | xargs)
+    alembic upgrade head
+fi
+
